backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
python-dotenv>=1.0.0

# Translation (Google Gemini)
google-generativeai>=0.8.0

# PDF Processing
pdfplumber>=0.11.0
pdf2docx>=0.5.8
python-docx>=1.1.0

# Data Processing
pandas>=2.2.0
openpyxl>=3.1.2
